+++ element/al_toggle.js
@@ -42,12 +42,16 @@ AlchemyToggle.setAttribute('value', null, function setValue(value) {
  *
  * @author   Jelle De Loecker <jelle@elevenways.be>
  * @since    0.1.0
- * @version  0.1.0
+ * @version  0.2.9
  */
 AlchemyToggle.setMethod(function introduced() {
 
 	this.addEventListener('click', e => {
 		this.value = !this.value;
+		this.emit('change', {
+			'bubbles'   : true,
+			'cancelable': true
+		});
 	});
 
 	let input = this.querySelector('input');
@@ -62,5 +66,9 @@ AlchemyToggle.setMethod(function introduced() {
 		}
 
 		this.value = input.checked;
+		this.emit('change', {
+			'bubbles'   : true,
+			'cancelable': true
+		});
 	});
 });