+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.9 (WIP)
 
 * Add support for new local Date/Time fields & Decimal fields
+* Add "hidden" field templates
 
 ## 0.2.8 (2023-10-15)