+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add "hidden" field templates
 * Fix `al-toggle` not firing a change event
 * Fixed `al-field` rendering to update content for non-value setter elements when a new value is assigned
+* Use `constraints` data from sibling fields when loading associated field data
 
 ## 0.2.8 (2023-10-15)